cloudflare-examples/cloudflare-airtable/requirements.txt
cloudflare==2.1.0
airtable-python-wrapper==0.11.2
